Seperated GUI from grunt work
